src/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "V.Guilherme",email = "v.guilherme.barreto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.118.2,<0.119.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic[email] (>=2.12.0,<3.0.0)",
    "validate-docbr (>=1.11.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
